# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Pal {
  id: number;
  name: string;
  nameTransliteration: string;
  meaning: string;
}

export interface Adhikaram {
  id: number;
  palId: number;
  name: string;
  nameTransliteration: string;
  meaning: string;
  section?: string;
  introduction?: string;
}

export interface Kural {
  number: number;
  adhikaramId: number;
  tamil: string;
  transliteration: string;
  commentary: string;
}

export type GetAdhikaramsParams = {
  palId?: number;
};

export type GetKuralsParams = {
  adhikaramId?: number;
};