# lib/api-zod/src/generated/types/adhikaram.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Adhikaram {
  id: number;
  palId: number;
  name: string;
  nameTransliteration: string;
  meaning: string;
  section?: string;
  introduction?: string;
}