# lib/api-zod/src/generated/types/kural.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Kural {
  number: number;
  adhikaramId: number;
  tamil: string;
  transliteration: string;
  commentary: string;
}