# lib/api-zod/src/generated/types/pal.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Pal {
  id: number;
  name: string;
  nameTransliteration: string;
  meaning: string;
}